[refactor] Move models, services, and mocks in dedicated folders
Fix #14
