+++ app/projects/projects.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { ProjectsService } from './projects.service'
-import { Group } from './shared/group';
+import { ProjectsService } from '../+services/projects.service'
+import { Group } from '../+models/group';
 import { ErrorService } from '../+services/error.service';
 
 @Component({
